Extrach ethernet packet body extractor into separate tool
